refactor: move repositories directory
